Tests to ensure optimal watches on zk.
ZK itself does not duplicate watches(ie with the same session watching
the same node multiple times does not cause the watch count to
increment)

But the node-zookeeper client seems to create multiple watchers and
callbacks for each watch on the same node. Hence would be useful to
avoid that.